refactor: Réorganisation de la structure des fichiers pour améliorer la lisibilité et la maintenance
