+++ src/components/pages/homePage/HomePage.tsx
@@ -1,7 +1,7 @@
 // Components
-import Banner from './sections/Banner.tsx';
-import About from './sections/About.tsx';
-import Skills from './sections/Skills.tsx';
+import Banner from './sections/Banner/Banner.tsx';
+import About from './sections/About/About.tsx';
+import Skills from './sections/Skills/Skills.tsx';
 
 /** Affiche la page d'accueil du site
  * @description - Contient les sections Banner, About et Skills.
